refactor(editor): block selection navigation 분리
