+++ front/e2e/editor-studio-state.spec.ts
@@ -900,6 +900,12 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toMatch(/const\s+isStableBlockHandleState\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+isStableBlockSelectionOverlayState\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveOuterBlockSelectionGesture\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+getTopLevelBlockIndexFromSelection\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+getTopLevelBlockPosition\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+getFirstEditableTextPositionInNode\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+getEditableTextPositionForTopLevelBlock\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+selectTopLevelBlockNode\s*=/)
+    expect(blockEditorEngineSource).not.toMatch(/const\s+isTabBlockSelectionEligible\s*=/)
     expect(blockEditorEngineSource).toContain('from "./blockHandleLayoutModel"')
     expect(blockEditorEngineSource).toContain('from "./blockDragModel"')
     expect(blockEditorEngineSource).toContain('from "./tableAxisDragModel"')
@@ -936,6 +942,13 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toContain("sourceElement.textContent?.trim().replace(/\\s+/g, \" \").slice(0, 220)")
     expect(blockSelectionModelSource).toContain("export const resolveOuterBlockSelectionGesture")
     expect(blockSelectionModelSource).toContain("export const resolveOuterListItemSelectionGesture")
+    expect(blockSelectionModelSource).toContain("export const getTopLevelBlockIndexFromSelection")
+    expect(blockSelectionModelSource).toContain("export const getTopLevelBlockPosition")
+    expect(blockSelectionModelSource).toContain("export const getFirstEditableTextPositionInNode")
+    expect(blockSelectionModelSource).toContain("export const getEditableTextPositionForTopLevelBlock")
+    expect(blockSelectionModelSource).toContain("export const selectTopLevelBlockNode")
+    expect(blockSelectionModelSource).toContain("export const isTabBlockSelectionEligible")
+    expect(blockEditorEngineSource.split("\n").length).toBeLessThan(10480)
     expect(blockHandleLayoutModelSource).toContain("export const resolveBlockHandleAnchorTop")
     expect(blockHandleLayoutModelSource).toContain("export const resolveThinBlockHandleAnchorTop")
     expect(blockHandleLayoutModelSource).toContain("export const resolveBlockHandleRailLayout")
