+++ front/src/components/editor/blockSelectionModel.ts
@@ -1,3 +1,10 @@
+import type { Editor as TiptapEditor } from "@tiptap/core"
+import type { Node as ProseMirrorNode } from "@tiptap/pm/model"
+import { NodeSelection } from "@tiptap/pm/state"
+import type { BlockEditorDoc } from "./serialization"
+import { shouldCenterBlockHandleForNode } from "./blockHandleLayoutModel"
+import { isTableSelectionActive } from "./tableStructureModel"
+
 export type TopLevelBlockHandleState = {
   visible: boolean
   kind: "top-level" | "list-item"
@@ -66,6 +73,102 @@ export const isStableBlockSelectionOverlayState = (
   isWithinBlockHandleEpsilon(prev.top, next.top) &&
   isWithinBlockHandleEpsilon(prev.width, next.width) &&
   isWithinBlockHandleEpsilon(prev.height, next.height)
+
+export const getTopLevelBlockIndexFromSelection = (editor: TiptapEditor) => {
+  const { selection } = editor.state
+  return Math.max(0, selection.$from.index(0))
+}
+
+export const getTopLevelBlockPosition = (editor: TiptapEditor, blockIndex: number) => {
+  const { doc } = editor.state
+  if (doc.childCount === 0) return 1
+  const clampedIndex = Math.max(0, Math.min(blockIndex, doc.childCount - 1))
+  let position = 1
+  for (let index = 0; index < clampedIndex; index += 1) {
+    position += doc.child(index).nodeSize
+  }
+  return position
+}
+
+export const getFirstEditableTextPositionInNode = (
+  node: ProseMirrorNode | null | undefined,
+  startPos: number
+): number | null => {
+  if (!node) return null
+  if (node.isTextblock) {
+    return startPos + 1
+  }
+
+  if (!node.childCount) {
+    return null
+  }
+
+  let childPos = startPos + 1
+  for (let index = 0; index < node.childCount; index += 1) {
+    const child = node.child(index)
+    const nested = getFirstEditableTextPositionInNode(child, childPos)
+    if (nested !== null) {
+      return nested
+    }
+    childPos += child.nodeSize
+  }
+
+  return null
+}
+
+export const getEditableTextPositionForTopLevelBlock = (editor: TiptapEditor, blockIndex: number) => {
+  const { doc } = editor.state
+  if (doc.childCount === 0) return null
+  const clampedIndex = Math.max(0, Math.min(blockIndex, doc.childCount - 1))
+  const topLevelBlock = doc.child(clampedIndex)
+  if (
+    ![
+      "paragraph",
+      "heading",
+      "blockquote",
+      "bulletList",
+      "orderedList",
+      "taskList",
+      "calloutBlock",
+      "toggleBlock",
+    ].includes(topLevelBlock.type.name)
+  ) {
+    return null
+  }
+
+  const blockPosition = getTopLevelBlockPosition(editor, clampedIndex)
+  return getFirstEditableTextPositionInNode(topLevelBlock, blockPosition)
+}
+
+const focusEditorViewWithoutScroll = (editor: TiptapEditor) => {
+  if (typeof window === "undefined") {
+    editor.view.focus()
+    return
+  }
+
+  const previousScrollX = window.scrollX
+  const previousScrollY = window.scrollY
+  editor.view.focus()
+  if (window.scrollX !== previousScrollX || window.scrollY !== previousScrollY) {
+    window.scrollTo(previousScrollX, previousScrollY)
+  }
+}
+
+export const selectTopLevelBlockNode = (editor: TiptapEditor, blockIndex: number) => {
+  const { doc, tr } = editor.state
+  if (doc.childCount === 0) return
+  const clampedIndex = Math.max(0, Math.min(blockIndex, doc.childCount - 1))
+  const position = getTopLevelBlockPosition(editor, clampedIndex)
+  const selection = NodeSelection.create(doc, position)
+  editor.view.dispatch(tr.setSelection(selection))
+  focusEditorViewWithoutScroll(editor)
+}
+
+export const isTabBlockSelectionEligible = (editor: TiptapEditor, blockIndex: number | null) => {
+  if (blockIndex === null || isTableSelectionActive(editor)) return false
+  const blocks = ((editor.getJSON() as BlockEditorDoc).content ?? []) as BlockEditorDoc[]
+  return shouldCenterBlockHandleForNode(blocks[blockIndex] ?? null)
+}
 
 const isSelectionPointerGesture = (event: BlockSelectionPointerEventLike) =>
   event.button === 0 &&
